ref: Outsource database functions
