+++ .gitignore
@@ -131,3 +131,4 @@ dmypy.json
 # Own
 config.ini
 data*
+!database.py
